My first commit

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	float f;
 	char d;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a long int: %lu	byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a char: %lu byte(s)\n", sizeof(d));
+	printf("Size of an int: %lu byte(s)\n", sizeof(a));
+	printf("Size of a long int: %lu	byte(s)\n", sizeof(b));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(c));
 	printf("Size of a foat: %lu byte(s) \n", (unsigned long)sizeof(f));
 	return (0);
 }
